requirements.txt
feedparser==6.0.12
supabase==2.27.3
python-dotenv==1.2.1
pydantic==2.12.0
requests==2.32.5
pytest==9.0.2
openai==2.17.0
tzdata==2025.1
